+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-29 16:30:53
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-05-29 18:39:21
+// Commit: fix(test): correct layout constraints and overflow in test
 
 class testComponentState {
-  final String commitRef = '1588';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-05-29T16:30:53+05:00');
+  final String commitRef = '1597';
+  final String lastMessage = 'fix(test): correct layout constraints and overflow in test';
+  final DateTime timestamp = DateTime.parse('2026-05-29T18:39:21+05:00');
 }